eks-pods/dashboard-svc/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.0
httpx==0.28.1
redis==5.2.1
websockets==14.1
psycopg[binary,pool]==3.2.13
PyJWT[crypto]==2.10.1
